+++ src/main/java/kitchenpos/application/dto/request/OrderTableRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto;
+package kitchenpos.application.dto.request;
 
 public class OrderTableRequest {
