+++ src/main/java/kitchenpos/ui/TableGroupRestController.java
@@ -1,7 +1,7 @@
 package kitchenpos.ui;
 
 import kitchenpos.application.TableGroupService;
-import kitchenpos.application.dto.TableGroupCreateRequest;
+import kitchenpos.application.dto.request.TableGroupCreateRequest;
 import kitchenpos.domain.TableGroup;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
